+++ docker/celestia/run-validator.sh
@@ -147,12 +147,12 @@ setup_private_validator() {
   dasel put -f "$CONFIG_DIR/config/config.toml" -t string -v '0.0.0.0:26660' instrumentation.prometheus_listen_addr
 
   # Adjusting for faster block times. Default values are in comments above
-  # Currently disabled
+  # Currently disabled to be closer to actual mocha
 #  # Numbers are derived by trial and error.
 #  # timeout_commit = "11s"
-#  dasel put -f "$CONFIG_DIR/config/config.toml" -t string -v '2000ms' consensus.timeout_commit
+#  dasel put -f "$CONFIG_DIR/config/config.toml" -t string -v '3000ms' consensus.timeout_commit
 #  # timeout_propose = "10s"
-#  dasel put -f "$CONFIG_DIR/config/config.toml" -t string -v '4000ms' consensus.timeout_propose
+#  dasel put -f "$CONFIG_DIR/config/config.toml" -t string -v '2000ms' consensus.timeout_propose
 #  # timeout_propose_delta = "500ms"
 #  dasel put -f "$CONFIG_DIR/config/config.toml" -t string -v '200ms' consensus.timeout_propose_delta
 #  # timeout_prevote = "1s"